+++ src/geoai3d/__init__.py
@@ -13,6 +13,7 @@ from geoai3d.crs import reproject, reproject_3d
 from geoai3d.features import geometric_features, multiscale_features
 from geoai3d.geometry import estimate_normals, remove_statistical_outliers
 from geoai3d.index import build_kdtree
+from geoai3d.intensity import normalize_intensity
 from geoai3d.io.dispatch import read_lidar
 from geoai3d.io.las import read_las, to_las
 from geoai3d.io.parquet import read_parquet, to_parquet
@@ -36,6 +37,7 @@ __all__ = [
     "geometric_features_stream",
     "geometric_features_tiled",
     "multiscale_features",
+    "normalize_intensity",
     "read_las",
     "read_lidar",
     "read_parquet",